requirements.txt
praw
python-dotenv
chromadb
openai
tiktoken
nltk
matplotlib
seaborn
pandas
scikit-learn
wordcloud
textblob
streamlit
plotly